Move from /zui into /features/call
