+++ src/features/call/components/Report/steps/LeftMessage.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react';
 
 import { ReportType } from '..';
 import ZUIButtonGroup from 'zui/components/ZUIButtonGroup';
-import messageIds from 'zui/l10n/messageIds';
 import { Msg, useMessages } from 'core/i18n';
+import messageIds from 'features/call/l10n/messageIds';
 import StepBase from './StepBase';
 
 type Props = {
@@ -11,33 +11,31 @@ type Props = {
   report: ReportType;
 };
 
-const OrganizerAction: FC<Props> = ({ onReportUpdate, report }) => {
+const LeftMessage: FC<Props> = ({ onReportUpdate, report }) => {
   const messages = useMessages(messageIds);
   return (
     <StepBase
       state="active"
-      title={
-        <Msg id={messageIds.report.steps.organizerAction.question.title} />
-      }
+      title={<Msg id={messageIds.report.steps.leftMessage.question.title} />}
     >
       <ZUIButtonGroup
         buttons={[
           {
-            label: messages.report.steps.organizerAction.question.yesButton(),
+            label: messages.report.steps.leftMessage.question.yesButton(),
             onClick: () =>
               onReportUpdate({
                 ...report,
-                organizerActionNeeded: true,
-                step: 'orgLog',
+                leftMessage: true,
+                step: 'orgAction',
               }),
           },
           {
-            label: messages.report.steps.organizerAction.question.noButton(),
+            label: messages.report.steps.leftMessage.question.noButton(),
             onClick: () =>
               onReportUpdate({
                 ...report,
-                organizerActionNeeded: false,
-                step: 'callerLog',
+                leftMessage: false,
+                step: 'orgAction',
               }),
           },
         ]}
@@ -48,4 +46,4 @@ const OrganizerAction: FC<Props> = ({ onReportUpdate, report }) => {
   );
 };
 
-export default OrganizerAction;
+export default LeftMessage;
