+++ src/features/call/components/Report/steps/OrganizerLog.tsx
@@ -5,7 +5,7 @@ import { ReportType } from '..';
 import ZUITextField from 'zui/components/ZUITextField';
 import ZUIButton from 'zui/components/ZUIButton';
 import { Msg, useMessages } from 'core/i18n';
-import messageIds from 'zui/l10n/messageIds';
+import messageIds from 'features/call/l10n/messageIds';
 import StepBase from './StepBase';
 import { ZetkinCallTarget } from 'features/call/types';
