+++ src/features/call/components/Report/steps/Summary.tsx
@@ -4,14 +4,14 @@ import StepBase from './StepBase';
 
 type Props = {
   onEdit: () => void;
-  subtitle?: JSX.Element;
   state: 'active' | 'failure' | 'success';
+  subtitle?: JSX.Element;
   title: JSX.Element;
 };
 
 const Summary: FC<Props> = ({ onEdit, subtitle, state, title }) => {
   return (
-    <StepBase onEdit={onEdit} subtitle={subtitle} state={state} title={title} />
+    <StepBase onEdit={onEdit} state={state} subtitle={subtitle} title={title} />
   );
 };
